+++ config/webpack.config.base.js
@@ -21,7 +21,8 @@ module.exports = {
         ],
     },
     output : {
-        libraryTarget: 'umd',
+        libraryTarget: 'umd', // For info, using `var` instead of `umd` generates a cleaner dist file
+        libraryExport: 'default', // This option removes the `.default` after the build (see https://github.com/webpack/webpack/issues/3929)
         library      : 'AutoNumeric',
         filename     : 'autoNumeric.js',
         path         : resolve('dist'),