+++ src/AutoNumeric.js
@@ -1,8 +1,8 @@
 /**
  *               AutoNumeric.js
  *
- * @version      4.5.4
- * @date         2019-01-05 UTC 07:40
+ * @version      4.5.7
+ * @date         2019-09-13 UTC 07:47
  *
  * @authors      Bob Knothe, Alexandre Bonneau
  * @contributors Sokolov Yura and others, cf. AUTHORS
@@ -113,7 +113,7 @@ export default class AutoNumeric {
         this._checkElement();
 
         // Store the additional attributes inside the AutoNumeric object
-        // Note: This variable is needed and not a duplicate of `initialValueOnKeydown` nor `valueOnFocus` since it serves a different purpose and has a different lifecycle
+        // Note: This variable is needed and not a duplicate of `initialValueOnFirstKeydown` nor `valueOnFocus` since it serves a different purpose and has a different lifecycle
         this.savedCancellableValue = null;
 
         // Initialize the undo/redo variables
@@ -909,7 +909,7 @@ export default class AutoNumeric {
      * @returns {string}
      */
     static version() {
-        return '4.5.4';
+        return '4.5.7';
     }
 
     /**
@@ -1106,7 +1106,7 @@ export default class AutoNumeric {
      * @private
      */
     static _isPreDefinedOptionValid(preDefinedOptionName) {
-        return AutoNumeric.predefinedOptions.hasOwnProperty(preDefinedOptionName);
+        return Object.prototype.hasOwnProperty.call(AutoNumeric.predefinedOptions, preDefinedOptionName);
     }
 
     /**
@@ -4176,19 +4176,19 @@ export default class AutoNumeric {
 
         if (!AutoNumericHelper.isNull(options.styleRules) &&
             !(AutoNumericHelper.isObject(options.styleRules) &&
-            ((options.styleRules.hasOwnProperty('positive') ||
-            options.styleRules.hasOwnProperty('negative') ||
-            options.styleRules.hasOwnProperty('ranges') ||
-            options.styleRules.hasOwnProperty('userDefined'))))) {
+            ((Object.prototype.hasOwnProperty.call(options.styleRules, 'positive') ||
+            Object.prototype.hasOwnProperty.call(options.styleRules, 'negative') ||
+            Object.prototype.hasOwnProperty.call(options.styleRules, 'ranges') ||
+            Object.prototype.hasOwnProperty.call(options.styleRules, 'userDefined'))))) {
             AutoNumericHelper.throwError(`The option 'styleRules' is invalid ; it should be a correctly structured object, with one or more 'positive', 'negative', 'ranges' or 'userDefined' attributes, [${options.styleRules}] given.`);
         }
 
         // Deeper tests of the `styleRules` object : Check that the callback, if defined, is a function
         if (!AutoNumericHelper.isNull(options.styleRules) &&
-            options.styleRules.hasOwnProperty('userDefined') &&
+            Object.prototype.hasOwnProperty.call(options.styleRules, 'userDefined') &&
             !AutoNumericHelper.isNull(options.styleRules.userDefined)) {
             options.styleRules.userDefined.forEach(rule => {
-                if (rule.hasOwnProperty('callback') && !AutoNumericHelper.isFunction(rule.callback)) {
+                if (Object.prototype.hasOwnProperty.call(rule, 'callback') && !AutoNumericHelper.isFunction(rule.callback)) {
                     AutoNumericHelper.throwError(`The callback defined in the \`userDefined\` attribute is not a function, ${typeof rule.callback} given.`);
                 }
             });
@@ -4613,7 +4613,7 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
 
         if (!this.isManagedByAutoNumeric(domElement)) {
             let showWarnings;
-            if (!AutoNumericHelper.isNull(options) && options.hasOwnProperty('showWarnings')) {
+            if (!AutoNumericHelper.isNull(options) && Object.prototype.hasOwnProperty.call(options, 'showWarnings')) {
                 showWarnings = options.showWarnings;
             } else {
                 showWarnings = true;
@@ -4837,7 +4837,7 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
     _runCallbacksFoundInTheSettingsObject() { //FIXME test this
         // Loops through the this.settings object (option array) to find the following
         for (const key in this.settings) {
-            if (this.settings.hasOwnProperty(key)) {
+            if (Object.prototype.hasOwnProperty.call(this.settings, key)) {
                 const value = this.settings[key];
 
                 if (typeof value === 'function') {
@@ -6403,11 +6403,10 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
         }
 
         this._updateEventKeyInfo(e);
-        this.initialValueOnKeydown = AutoNumericHelper.getElementValue(e.target); // This is needed in `onKeyup()` to check if the value as changed during the key press
-        this.initialRawValueOnKeydown = this.rawValue;
 
         this.keydownEventCounter += 1; // Every time the keydown event is caught, increment the counter to keep track if the key is continuously pressed
         if (this.keydownEventCounter === 1) {
+            this.initialValueOnFirstKeydown = AutoNumericHelper.getElementValue(e.target); // This is needed in `onKeyup()` to check if the value as changed during the key press
             this.initialRawValueOnFirstKeydown = this.rawValue;
         }
 
@@ -6581,7 +6580,6 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
      */
     _onKeyup(e) {
         this.isEditing = false;
-        const multipleKeydownEvents = this.keydownEventCounter > 1;
         this.keydownEventCounter = 0; // Reset the keydown events counter
 
         if (this.formulaMode) {
@@ -6647,7 +6645,7 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
         const skip = this._processNonPrintableKeysAndShortcuts(e);
         delete this.valuePartsBeforePaste;
         const targetValue = AutoNumericHelper.getElementValue(e.target);
-        if (skip || !multipleKeydownEvents && targetValue === '') { // When the user keeps pressing the backspace or delete key, and end up deleting the entire input text, then the target value is equal to '', but we shouldn't `return` without first testing what was the initial value when the user started pressing any key (cf. issue #621)
+        if (skip || targetValue === '' && this.initialValueOnFirstKeydown === '') { // If the user enters skippable keys, or keeps deleting/backspacing into the empty input, no 'formatted' event are sent (cf. issue #621)
             return;
         }
 
@@ -6680,12 +6678,11 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
         this._saveRawValueForAndroid();
 
         // If the input value has changed during the key press event chain, an event is sent to alert that a formatting has been done (cf. Issue #187)
-        if ((multipleKeydownEvents && targetValue !== this.initialRawValueOnFirstKeydown) || // If multiple keydown events are detected, then we need to check the rawValue saved on the very first event
-            targetValue !== this.initialValueOnKeydown) {
+        if (targetValue !== this.initialValueOnFirstKeydown) { // Checking the value that were saved after the very first keydown event ensure that the 'formatted' event will be sent even if the user continuously press the Delete/Backspace key
             this._triggerEvent(AutoNumeric.events.formatted, e.target, {
-                oldValue   : this.initialValueOnKeydown,
+                oldValue   : this.initialValueOnFirstKeydown,
                 newValue   : targetValue,
-                oldRawValue: this.initialRawValueOnKeydown,
+                oldRawValue: this.initialRawValueOnFirstKeydown,
                 newRawValue: this.rawValue,
                 isPristine : this.isPristine(false),
                 error      : null,
@@ -7943,7 +7940,7 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
      */
     _transformOptionsValuesToDefaultTypes() {
         for (const key in this.settings) {
-            if (this.settings.hasOwnProperty(key)) {
+            if (Object.prototype.hasOwnProperty.call(this.settings, key)) {
                 const value = this.settings[key];
 
                 // Convert the strings 'true' and 'false' to booleans
@@ -8069,13 +8066,13 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
         };
 
         for (const option in options) {
-            if (options.hasOwnProperty(option)) {
+            if (Object.prototype.hasOwnProperty.call(options, option)) {
                 if (oldOptionsConverter[option] === true) {
                     // If the option is a 'new' option, we continue looping
                     continue;
                 }
 
-                if (oldOptionsConverter.hasOwnProperty(option)) {
+                if (Object.prototype.hasOwnProperty.call(oldOptionsConverter, option)) {
                     // Else we have an 'old' option name
                     AutoNumericHelper.warning(`You are using the deprecated option name '${option}'. Please use '${oldOptionsConverter[option]}' instead from now on. The old option name will be dropped very soon™.`, true);
 
@@ -8218,7 +8215,7 @@ To solve that, you'd need to either set \`decimalPlacesRawValue\` to \`null\`, o
     static _toNumericValue(value, settings) {
         //XXX Note; this function is static since we need to pass a `settings` object when calling the static `AutoNumeric.format()` method
         let result;
-        if (AutoNumericHelper.isNumber(Number(value))) {
+        if (AutoNumericHelper.isNumber(Number(value))) { // if (settings.decimalCharacter === '.' && AutoNumericHelper.isNumber(Number(value))) {
             // The value has either already been stripped, or a 'real' javascript number is passed as a parameter
             result = AutoNumericHelper.scientificToDecimal(value);
         } else {
@@ -9263,13 +9260,13 @@ AutoNumeric.multiple = (arg1, initialValue = null, options = null) => {
     if (AutoNumericHelper.isString(arg1)) {
         arg1 = [... document.querySelectorAll(arg1)]; // Convert a NodeList to an Array (cf. http://stackoverflow.com/a/37297292/2834898)
     } else if (AutoNumericHelper.isObject(arg1)) {
-        if (!arg1.hasOwnProperty('rootElement')) {
+        if (!Object.prototype.hasOwnProperty.call(arg1, 'rootElement')) {
             AutoNumericHelper.throwError(`The object passed to the 'multiple' function is invalid ; no 'rootElement' attribute found.`);
         }
 
         // Retrieve the DOM element list from the given <form> element
         const elements = [... arg1.rootElement.querySelectorAll('input')];
-        if (arg1.hasOwnProperty('exclude')) {
+        if (Object.prototype.hasOwnProperty.call(arg1, 'exclude')) {
             if (!Array.isArray(arg1.exclude)) {
                 AutoNumericHelper.throwError(`The 'exclude' array passed to the 'multiple' function is invalid.`);
             }
@@ -9289,7 +9286,7 @@ AutoNumeric.multiple = (arg1, initialValue = null, options = null) => {
             showWarnings = options.showWarnings;
         }
 
-        AutoNumericHelper.warning(`No valid DOM elements were given hence no AutoNumeric object were instantiated.`, showWarnings);
+        AutoNumericHelper.warning(`No valid DOM elements were given hence no AutoNumeric objects were instantiated.`, showWarnings);
 
         return [];
     }