+++ test/unit/karma.conf.js
@@ -18,7 +18,7 @@ module.exports = function(config) {
 
         // list of files / patterns to load in the browser
         files: [
-            '../../node_modules/babel-polyfill/dist/polyfill.js',
+            '../../node_modules/@babel/polyfill/dist/polyfill.js',
             './tests.webpack.js',
         ],